+++ src/editors/MiscEditor.h
@@ -1,5 +1,5 @@
 /*╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌
- * editors/MiscEditor.h is part of Brewtarget, and is copyright the following authors 2009-2024:
+ * editors/MiscEditor.h is part of Brewtarget, and is copyright the following authors 2009-2025:
  *   • Jeff Bailey <skydvr38@verizon.net>
  *   • Matt Young <mfsy@yahoo.com>
  *   • Mik Firestone <mikfire@gmail.com>
@@ -29,7 +29,7 @@
 #include "editors/EditorBase.h"
 #include "model/Misc.h"
 
-#define MiscEditorOptions EditorBaseOptions{ .nameTab = true, .idDisplay = true }
+#define MiscEditorOptions EditorBaseOptions{ .nameTab = true, .idDisplay = true, .numRecipesUsing = true }
 /*!
  * \class MiscEditor
  *